Move helper functions into own module
Helper functions are now inside utils.py, and are just imported from
there.
